+++ QuickSort/QuickSort.vcxproj.filters
@@ -15,7 +15,10 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="源.cpp">
+    <ClCompile Include="快排.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
+    <ClCompile Include="快选.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
   </ItemGroup>
