+++ QuickSort/x64/Debug/QuickSort.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>C:\Users\JYS\Desktop\C\C++algorithm\QuickSort\x64\Debug\QuickSort.exe</FullPath>
+      <FullPath>D:\tool\VS\project\C++algroithm\QuickSort\x64\Debug\QuickSort.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
