+++ QuickSort/x64/Debug/QuickSort.tlog/QuickSort.lastbuildstate
@@ -1,2 +1,2 @@
 PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.33.31629:TargetPlatformVersion=10.0.19041.0:
-Debug|x64|C:\Users\JYS\Desktop\C\C++algorithm\QuickSort\|
+Debug|x64|D:\tool\VS\project\C++algroithm\QuickSort\|
