+++ QuickSort/x64/Debug/QuickSort.vcxproj.FileListAbsolute.txt
@@ -1,1 +1,2 @@
 C:\Users\JYS\Desktop\C\C++algorithm\QuickSort\x64\Debug\QuickSort.exe
+D:\tool\VS\project\C++algroithm\QuickSort\x64\Debug\QuickSort.exe
